+++ src/item/data/ArmorDataModel.ts
@@ -7,6 +7,7 @@
  */
 
 import EquipmentDataModel from '@/item/data/EquipmentDataModel';
+import { ContainedItemQuality } from '@/item/data/WeaponDataModel';
 
 export default abstract class ArmorDataModel extends EquipmentDataModel {
 	/**
@@ -19,6 +20,11 @@ export default abstract class ArmorDataModel extends EquipmentDataModel {
 	 */
 	abstract soak: number;
 
+	/**
+	 * Item qualities for the armor.
+	 */
+	abstract qualities: ContainedItemQuality[];
+
 	static override defineSchema() {
 		const fields = foundry.data.fields;
 
@@ -26,6 +32,14 @@ export default abstract class ArmorDataModel extends EquipmentDataModel {
 			...super.defineSchema(),
 			defense: new fields.NumberField({ integer: true, initial: 0 }),
 			soak: new fields.NumberField({ integer: true, initial: 0 }),
+			qualities: new fields.ArrayField(
+				new fields.SchemaField({
+					name: new fields.StringField(),
+					description: new fields.HTMLField(),
+					isRated: new fields.BooleanField({ initial: true }),
+					rating: new fields.NumberField({ initial: 1, integer: true }),
+				}),
+			),
 		};
 	}
 }