+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Alpha 0.1.1
 ### Updates:
 - Disallowed adding ActiveEffects to Item Qualities until a better solution is implemented for referencing or embedding them in items.
+- [#25](https://github.com/Mezryss/FVTT-Genesys/issues/25) Allow adding item qualities to Armor.
 ### Bug Fixes:
 - Fix [#22](https://github.com/Mezryss/FVTT-Genesys/issues/22): In skills tab, context menu can appear below XP Container and other skill categories.
 - Fix [#27](https://github.com/Mezryss/FVTT-Genesys/issues/27): Disallow archetype removal if XP has changed since Archetype was applied.